requirements.txt
<<<<<<< HEAD
streamlit==1.32.0
pandas
numpy
groq
python-dotenv
PyPDF2
docx2txt
plotly
scikit-learn
Pillow


# SharePoint / Microsoft Graph
msal==1.26.0
requests==2.31.0
python-dotenv==1.0.1

=======
streamlit==1.31.0
groq>=0.11.0
pandas==2.1.4
PyPDF2==3.0.1
docx2txt==0.8
plotly==5.18.0
python-docx==1.1.0
openpyxl==3.1.2
pytesseract==0.3.10
Pillow==10.1.0
scikit-learn==1.3.2
Office365-REST-Python-Client==2.5.3
msal==1.24.0
>>>>>>> dev-branch
